requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-jazzmin==2.6.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jwt==1.3.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
